{{cookiecutter.project_slug}}/requirements.txt
# Template Core Requirements

fastapi==0.75.0
uvicorn==0.17.5
python-dotenv==0.20.0

# Template Optional Requirements
httpx==0.22.0

# Custom Requirements